# Conflicts:
#	kernel/src/main/java/com/flyemu/share/entity/basic/Product.java
#	kernel/src/main/resources/application.yml
#	merchant/merchant-front/src/views/basic/ProductForm.vue
#	merchant/merchant-front/src/views/inventory/CostAdjustmentForm.vue
#	merchant/merchant-front/src/views/inventory/InventoryTransferForm.vue
#	merchant/merchant-front/src/views/inventory/OtherInboundForm.vue
#	merchant/merchant-front/src/views/inventory/OtherOutboundForm.vue
#	merchant/merchant-front/src/views/purchase/PurchaseInboundForm.vue
#	merchant/merchant-front/src/views/purchase/PurchaseOrderForm.vue
#	merchant/merchant-front/src/views/purchase/PurchaseReturnForm.vue
#	merchant/merchant-front/src/views/sales/SalesOrderForm.vue
#	merchant/merchant-front/src/views/sales/SalesOutboundForm.vue
#	merchant/merchant-front/src/views/sales/SalesReturnForm.vue

+++ kernel/src/main/java/com/flyemu/share/dto/purchase/PurchaseReportItemDto.java
@@ -75,12 +75,12 @@ public class PurchaseReportItemDto {
     /**
      * 采购数量
      */
-    private Double secondaryQuantity;
+    private BigDecimal secondaryQuantity;
 
     /**
      * 采购单价
      */
-    private Double secondaryPrice;
+    private BigDecimal secondaryPrice;
 
     /**
      * 采购金额
